refactor: allocation tracker refactor into partitionable_devices.go
